Revise following reviewers' comment.
Reviewer #4 found that the novel aspects of the paper were hard to
distinguish, so we rewrite the conclusion to make them more obvious.

Also, the same reviewer apparently believed that cross-species mappings
were previously already available in some kind of downloadable files and
that we merely converted that file from one format to another, so we add
a sentence in the section about the drawbacks of cross-references to
make it clearer that the mappings were not, in fact, already available
in any other form than cross-references embedded within the ontology.