Add back removed slides about the previous system.
We are now aiming for a 25-min talk, so we can afford to put back the
previously removed slides that expand about the limitatons of the
previous integration pipeline (but put the slide with the overview
first, not last).

Also move the slide about the ROBOT implementation back from the
supplementary slides to the main presentation.